% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_example_gdb.R
\name{create_example_gdb}
\alias{create_example_gdb}
\title{Generate an example gdb.}
\usage{
create_example_gdb(filepath = NULL)
}
\arguments{
\item{filepath}{Filepath to write example gdb to.
By default, the gdb is stored as a temporary file, which will be deleted at the
end of the R session.}
}
\value{
An object of class `gdb`: a connection to an RVAT gdb file.
}
\description{
Generates an example gdb.
By default it will be stored in a temporary file which will be deleted at the end of the R session.
Alternatively, the `filepath` parameter can be set to store the gdb in a location of choice.
}
\examples{
# generate example gdb
gdb <- create_example_gdb()

}
