% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GT.R
\docType{data}
\name{GT}
\alias{GT}
\title{Example genoMatrix object}
\format{
An example `[rvat::genoMatrix]` object.
Contains simulated genotype data of 25,000 individuals and 35 moderate/high impact variants in the ALS gene \emph{SOD1}
}
\usage{
GT
}
\description{
Example genoMatrix object
}
\keyword{datasets}
