% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTsmall.R
\docType{data}
\name{GTsmall}
\alias{GTsmall}
\title{Example small genoMatrix object}
\format{
An example `[rvat::genoMatrix]` object.
Contains simulated genotype data of 5,000 individuals and 35 moderate/high impact variants in the ALS gene \emph{SOD1}
}
\usage{
GTsmall
}
\description{
This object includes 5,000 randomly samples from the `[GT]` example data.
}
\keyword{datasets}
