% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvat_example.R
\name{rvat_example}
\alias{rvat_example}
\title{Get path(s) to rvat example data}
\usage{
rvat_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
This function retrieves the file paths of example data, including an example vcf-file
and gdb.
}
\examples{
rvat_example()
rvat_example("rvatData.gdb")
rvat_example("rvatData.vcf.gz")
}
