% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvbresults.R
\docType{data}
\name{rvbresults}
\alias{rvbresults}
\title{Example rvb results}
\format{
An example `[rvat::rvbResult]` object.
}
\usage{
rvbresults
}
\description{
This object includes simulated rare variant burden results including various statistical tests and variant sets.
}
\keyword{datasets}
